Final Project
Here are 3 code segments. The TwitterTry folder contains the Twitter4j
API and TwitterTry.java. This java class is used to query twitters and
process the returned information into a CSV file, which can be used as
the input for classifier. The logistic.py is used to take the input
from Twitter data and Bitcoin price data to do the classification by
logistic regression. The svm.py is used to take the input from the
dataset and do the classification by SVM. The final report is also
attached.